Remove _asset_list_config import in fv3config.config.derive
Needed to avoid circular import when trying to get coupler_res in
derive.py.

* Splits _asset_list into:

1. general I/O stuff
2. fv3config specific stuff (diag_table, namelist, etc)

* move the inital conditions specific logic out of "datastore".
* move datastore.ensure_exists to filesystem

refactor get_patch_file_assets

check in file i forgot
